requirements.txt
mido==1.2.9
pyserial==3.4
python-rtmidi==1.2.1
